Minor: rename things for consistency and vue style
