+++ assets/vue/router/links.js
@@ -13,22 +13,22 @@ export default {
     {
       name: 'CreateLink',
       path: 'create',
-      component: () => import('../views/links/CreateLink.vue')
+      component: () => import('../views/links/LinksCreate.vue')
     },
     {
       name: 'UpdateLink',
       path: 'edit/:id',
-      component: () => import('../views/links/UpdateLink.vue')
+      component: () => import('../views/links/LinksUpdate.vue')
     },
     {
       name: 'CreateLinkCategory',
       path: 'create_link_category',
-      component: () => import('../views/links/CreateLinkCategory.vue')
+      component: () => import('../views/links/LinksCategoryCreate.vue')
     },
     {
       name: 'UpdateLinkCategory',
       path: 'update_link_category/:id',
-      component: () => import('../views/links/UpdateLinkCategory.vue')
+      component: () => import('../views/links/LinksCategoryUpdate.vue')
     },
   ]
 };
