+++ assets/vue/views/links/LinksCreate.vue
